CI/Bats: disable the selftest_timedpass_fork test for the silent run
This test was added by commit 8c1466a1abea002f6eac023a390d546021a3d306
and is part of the standard '@positive' group of selftests. It appears
to pass just fine under normal circumstances, but we have observed it to
fail in:
- ASAN builds
- More than 10 threads

This appears to trigger some race condition inside libasan that causes
the child processes to remain behind. When that happens, Bats hangs.

This problem does not affect the "selftest_freeze_fork" bats test
because it runs with -n4 or fewer CPUs. Plus, in the freezing case, the
framework will kill those child processes, so Bats won't hang.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>