FW/Tests: refactor to allow SandstoneTestSet to keep per-test configuration
We were keeping the test configuration on a map, which meant the
configuration was unique per test name, instead of per test entry. This
commit merges the two collections into a vector of test_cfg_info.

We no longer need to keep the flag indicating if a test is enabled or
disabled: it is disabled if it is not on the test list at all. This
works because of the order of processing in main(): we first enable all
tests that need to be enabled (which is all if none were specified on
the command-line), then disable those the user asked to disable.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>