forkfd: Detect when waitid() returns with a reason not SIGCHLD
This can happen on Linux if we are ptracing the process, because
ptrace(2) uses the wait() mechanism to get new events from the traced
process. I don't know if anyone is using forkfd() to debug from the
parent (it should be possible), but I expect that they use the wait
functions directly instead of forkfd_wait() for the debugging work.

The problem is when something else traces our child processes. When a
buggy debugger detaches from our process, waitid returns without setting
si_signo (probably because SIGCHLD was never delivered). strace shows:

 waitid(P_PIDFD, 19, {}, WNOHANG|WEXITED, {...}) = 0

And the child process never changes state.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>