FW: ensure we handle tests that log_xxx() but EXIT_SUCCESS in cleanup()
Similar to the previous two commits, but now for the cleanup()
function. Unlike the last two commits, we didn't handle log_error() in
the clean up function either.

Drive-by fix running the cleanup in the case of strict runtime. Some
tests perform testing during clean up, so it's not optional.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>