Reverting dependency on libboost-all-dev in CI
Reverting PR #555 which modified dependency on Boost from`libboost-dev` to `libboost-all-dev`, which was needed for command lineparser refactoring. Initial round of review concluded that we'd like tokeep a header-only dependency on Boost.Also, `libboost-all-dev` introduces a large number of additional packageswhich cause a bit of problems in CI due to broken dependency chainsin Debian Sid.

Signed-off-by: Lukasz Tuz <lukasz.tuz@intel.com>