FW/Logging: refuse to overwrite time-based log files
When the user does not provide us with a name for the log file, we
generate one based on the current univesal time. This change makes
OpenDCDiag refuse to overwrite such an existing file.

The file name contains the current timestamp down to microseconds, so
it's extremely unlikely one *can* match an existing file anyway.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>