Framework extension to log a message without a use of logging facility
PTS console (as passed with PTS environment variable) is used to log
all the messages.
Default implementation to print (xbyak) code and to print buffer (as
bytes) included. Script print_asm.sh to parse the binary to assembly
form not included.
Feature to log the code to stderr (handled by the framework) added to
allow logging assembly of JITed code on a failure.

Signed-off-by: Jarek Poswiata <jaroslaw.poswiata@intel.com>