FW: make OpenDCDiag ignore SIGBUS containing MCEs if --ignore-mce-errors
In debug builds, we also ignore SIGTRAP (Breakpoints), which allows this
functionality to be tested. There's no way to send the required
`si_code` for SIGBUS from userspace - only the OS kernel is allowed to
set those values.

[ChangeLog][Framework] The --ignore-mce-errors will also cause the tool
to ignore SIGBUS errors sent by the Linux kernel with signal codes
indicating they were caused by MCEs. If those are detected, the test
will terminate the test and mark it as skipped.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>