FW/child_debug: fix returning from the signal handler if socket==-1
This isn't supposed to happen, because we don't install the signal
handler in the first place if we don't have both ends of the socket
pair. But in case we do, the correct thing to do is to repeat the same
suicide protocol.

Also fixing the detection of when to do so: we can detect a kernel-
originated signal with si_code >= 0.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>