FW/Win32: don't call SetThreadIdealProcessorEx under WINE
WINE doesn't implement that and prints a warning that you used it. That
used to be a "fixme" (which we can and do disable with the `WINEDEBUG`
environment variable), but recent versions have changed it to something
that always goes to stderr. That breaks our unit tests.

So instead let's detect it and avoid making such a call.

Ideally we'd just do `#ifndef NDEBUG` and enable it for debug mode,
but we can't build for Windows under debug mode due to limitations of
the file format relating to debug symbols (see commit
e9c9e294a6945c95173d920130a3ce45274ac64f). And we can't add to the
CXXFLAGS / cpp_flags when doing cross-compilation due to (intentional)
limitations in Meson.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>