FW: implement TEST_QUALITY_OPTIONAL
This is meant for tests that are production-quality, but for whatever
reason we don't want to have enabled by default. If the test is
explicitly named, it will get run (it won't match wildcards).

A good case for this may be an obsolete test, which may still be useful
in some circumstances, like for an A-B comparison with the updated
version.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>