CI: set $XDG_RUNTIME_DIR
During one of our tests under WINE, we're getting the error message:
```
error: XDG_RUNTIME_DIR is invalid or not set in the environment.
```

This is coming from libwayland-client, probably through Gtk as a
dependency of WINE.
https://gitlab.freedesktop.org/wayland/wayland/-/blob/1.23.1/src/wayland-client.c?ref_type=tags#L1154-1160

It began to be a problem with 3432e1190e343d71a516ceadedb6f31cfef62e63
(PR #587).

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>