FW: fix mismatch between - and _ in the AMX feature names
The GCC feature names (for `-m` and `__attribute__((target(xxx)))`) use
a dash, but the identifiers, by necessity, use an underscore. This
caused the XSAVE requirement "amx-tile" to not match feature
"amx_tile". The same problem would have happened for "sse4.1" and
"sse4.2", but there's no XSAVE feature for those.

That caused the script to emit:
```c++
// List of features requiring XSave_AmxState
static const uint64_t XSaveReq_AmxState = 0
        | cpu_feature_amx_bf16
        | cpu_feature_amx_int8
        | cpu_feature_amx_fp16;
```

As a result, a test that only required `amx_tile` would *not* get
disabled if we couldn't enable the XSAVE state via the Linux
`arch_prctl(2)` API or if the OS did not enable it with XGETBV.

Now it emits:
```c++
// List of features requiring XSave_AmxState
static const uint64_t XSaveReq_AmxState = 0
        | cpu_feature_amx_bf16
        | cpu_feature_amx_tile
        | cpu_feature_amx_int8
        | cpu_feature_amx_fp16;
```

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>