FW/child_debug: identify #VE and #CP fault names on x86
And fix the name of interrupt 19 (SIMD Floating-Point Exception) to
`#XM`. The name `XF` comes from the Linux kernel constant `X86_TRAP_XF`,
but that doesn't match the Intel manual (SDM Volume 3, Chapter 7).

`#VE` is "Virtualization Exception", which we are unlikely to get.

`#CP` is "Control Protection Exception", which we can get in userspace.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>