Self tests: Run with threads from two different cores.
Slicing would not separate threads from the same core into different
slices. When running the test on real hardware (and topology), -n2 and
--max-cores-per-slice=1 do not result in two separate slices and the
test will not find the main thread for the second slice in the output
and, thus, will fail. --cpuset=t0 ensures that only 1 thread from a core
is taken and will work if the topology has more than one core.

Signed-off-by: Arzhan Kinzhalin <arzhan.i.kinzhalin@intel.com>