Framework: Add `TestObsolete` skip category.
There are multiple reasons one may want to remove or make obsolete an existing content.
This commit purpose is to give means to developer to leave message for the user to explain what happened to the test.
We can not rely on release notes on this matter.

Making the obsolete test quality to `SKIP` is not an answer either as we do not leave any message when skipping due to test quality.

[ChangeLog] Added new runtime skip category called `TestObsolete`.

Signed-off-by: Dawid Sabat <dawid.sabat@intel.com>