FW: Fix empty messages in main threads when a test was skipped.
Adjust expected size by the size of the init skip message for main
threads which run init(). There's still a small potential issue with the
same code potentially printing a different skip message, but it needs to
be addressed separately and is, in fact, an indication of non-SMP
machine or something badly broken.

Before:

  result: skip
  skip-category: CpuTopologyIssue
  skip-reason: '2 threads on the same core are required'
  time-at-end:   { elapsed:      3.999, now: !!timestamp '2025-05-16T18:39:26Z' }
  test-runtime: 2.678
  threads:
  - thread: main
    messages:
- test: mce_check

After:

  result: skip
  skip-category: CpuTopologyIssue
  skip-reason: '2 threads on the same core are required'
  time-at-end:   { elapsed:      4.000, now: !!timestamp '2025-05-16T18:38:44Z' }
  test-runtime: 2.515
- test: mce_check

Signed-off-by: Arzhan Kinzhalin <arzhan.i.kinzhalin@intel.com>