FW: Add --include-optional command-line option.
This option will include the tests with quality level
TEST_QUALITY_OPTIONAL into the default run (without explicitly naming
all the optional tests).

It is a convenience option and in general would behave similarly to
--quality 1. However, it would not override lower quality levels if they
were specified. For example, in case of --beta --include-optional, the
outcome would still be --beta, unlike --quality 1 which would force the
quality level of TEST_QUALITY_OPTIONAL.

Signed-off-by: Arzhan Kinzhalin <arzhan.i.kinzhalin@intel.com>