Revert "FW: reorganize current_max_loop_count assignment"
This reverts commit bcb56fc858d28c08da11a9730497e46c57f607d7.

The change being reverted has an unintended consequence of increasing
runtime for the tests that specify .fracture_loop_count = -1 (no
fracture) and not honoring command-line-specified --max-test-loop-count
value. While exactly the intention of the previous change, it is
unacceptable to have the test hard-coded, unflexible logic. Thus
reverting it back. The tests must honor --max-test-loop-count on the
command-line if they can.

Signed-off-by: Arzhan Kinzhalin <arzhan.i.kinzhalin@intel.com>