FW: update the test planning to take modules and NUMA nodes into account
This was the objective all along when I wrote this code in 2023, but we
didn't have the information populated to achieve it. Therefore, it only
made a token attempt by checking if the core count in the package was a
multiple of 24 and by trying to keep things as a multiple of 4. This is
no longer needed.

Instead, the rules now are:

1. (unchanged) if we have 32 cores are under test or less, then no slicing
   happens, otherwise we slice at the package/socket level
2. (unchanged) if the package has 32 cores or less, the `isolate_socket` mode
   equals `heuristic`
3. otherwise, we split at the NUMA node boundaries
4. if the NUMA node has more than 32 cores, we subdivide again on roughly
   equally-sized slices
5. we keep all cores and threads of a given module inside of one slice

Keeping rules 1 and 2 unchanged means most older systems are unchanged
too[^1]. The first system with more than 32 cores were the double-die Cascade
and Cooper Lakes and later the Ice Lake on the same die, for which were
splitting at exactly half but now will split at the NUMA boundary (if any).

Splitting at NUMA node boundaries changes more recent processors with 3 nodes
or more. For example, Granite Rapids has a part with 120, which was being
split into 5x24-core slices and crossing NUMA boundaries; now it's split 6
ways.

Ideally we'd split at the tile level too, but I don't have access to any
system that reports tiles. Until I do, I can't tell whether it should sort
above the NUMA node or below, in splitting.

[ChangeLog][Framework] Updated the system-slicing rules to take the NUMA
node and module topology into account. Now the tool will preferably
split at NUMA node boundaries when those are discovered, which should
help improve performance on systems with more than 32 cores and an
uneven number of NUMA nodes, such as those found on Xeon 6P parts
(Granite Rapids). The tool will also keep all cores of a given module
in the same slice, which should improve Xeon 6E systems (Sierra Forest).

[^1]: except for the re-sorting of threads according to their NUMA node IDs,
which may affect timing as well as stack and heap pointers, but does not
affect the random generator nor slicing.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>