'device_type' build option. (#624)
* Reformatting build files with a Meson linter.

Signed-off-by: Lukasz Tuz <lukasz.tuz@intel.com>

* Adding the 'device_type' build option

Commit adds a device_type option to the OpenDCDiag. This option allows for specifying the target device during the build (currently does nothing and defaults to 'cpu').

In addition to adding this option to the 'meson_option.txt' file, a variable with the same name is added to the framework. In addition, the framework now checks if all variables it expects to be set are indeed created -- including the device_type.

Signed-off-by: Lukasz Tuz <lukasz.tuz@intel.com>

* kvm content built only for cpu device.

Signed-off-by: Łukasz Tuz <lukasz.tuz@intel.com>

* Fixes and cleanups.

Signed-off-by: Łukasz Tuz <lukasz.tuz@intel.com>

---------

Signed-off-by: Lukasz Tuz <lukasz.tuz@intel.com>
Signed-off-by: Łukasz Tuz <lukasz.tuz@intel.com>