FW: bail out early with an error message if we can't get our affinity
This is not likely to ever happen, assuming our implementations of
`ambient_logical_processor_set()` are correct (the Darwin one isn't! it
fakes the result). But in case it does, print an error to the user
instead of crashing on some null-pointer dereference or accessing
elements in a zero-length vector.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>