FW/child_debug: rewrite the GDB communication to handle large outputs
This code has been mostly unmodified since I wrote it in 2021, including
the "###" comment indicating that "gdb writes only a handful of bytes".

Unfortunately, that's not really true. On attaching to a running
process, the GDB preamble can actually be quite long and overflow our
4096-byte buffer. GDB prints the "[New LWP xxxxxx]" line for each thread
it detects (of which we can easily see more than 256 these days), the
current frame, and, on some systems, a long list of system packages that
it thinks are missing. The initial frame after attaching is usually
`syscall` or something equivalent inside of libc, but if our Python
script is successful at finding the actual crashing frame, its name may
be arbitrarily long. For example, I can see a 3958-character-long C++
function name from Eigen in our binaries.

This commit removes any limitation by reading straight into the
heap-allocated `std::string`, which we grow as needed.

Fixes #707.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>