FW: Expose test_requires_smt flag to tests
Why is this required?

Many tests in Sandstone need hyperthreading(smt) to be effective.
These tests need to write extra code to detect if there is a sibling
thread in the same core(HT). If no sibling thread present, they skip.
This flag will ease test writer's job where they can specify this flag
in test details and framework will skip this test if there are no hyperthreads.
If there is "atleast" 1 hyperthread, the test runs.

Signed-off-by: dilipshivaraju <dilip.shivaraju@intel.com>