CI/Bats: attempt to deal with multi-core module systems
The rules we had were fine so long as there was only one core per module
for the first few cores in the system. That's been true for what we've
tested in GitHub, in our internal CI, and even our laptops, but is not
the case for Intel Xeon 6E, with its quad-core modules. This commit
attempts to with that.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>