FW/Logging: fix crash when `find_package_id()` can't find the pkg
Only observed on Windows because one of the two code paths didn't call
`package_for_id()` to create the package entry — and unfortunately the
other code path is what GitHub Actions ran. Instead of forcing it
created on Windows (where we can't read the PPIN anyway), change the
code so we can cope with a null pointer return from `find_package_id()`.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>