FW/Options: add `--timeout-kill` to set the time before sending SIGKILL
This value is currently hardcoded at 20s. I don't expect anyone will
ever need to change the default, but I need it to test the framework's
ability to send `SIGKILL` to a process that refuses to exit after
`SIGQUIT`.

The time to wait after the `SIGKILL` remains hardcoded at 20s.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>