FW: fix race condition when child exits right as we detect its timeout
[ChangeLog][Framework] Fixed a race condition bug that caused the
framework to mis-identify a test that ran over the time limit but exited
cleanly right after that as a `result: fail` instead of either `result:
pass` (albeit late) or `result: timed out`.

This happened because the `TestResult::TimedOut` was assigned on Unix
systems by the child having exited due to a SIGQUIT signal, but if the
child exited before our `kill(2)` call, we would assign
`TestResult::Pass`. However, the framework does a `log_message()` of an
error, which sets the main thread state to failed, and thus causes the
entire test to be marked as having failed.

Before the fix:
```yaml
- test: selftest_freeze_exit_on_termination
  time-at-start: { elapsed:      0.000, now: !!timestamp '2025-07-22T21:20:07Z' }
  result: fail
  time-at-end:   { elapsed:   1073.004, now: !!timestamp '2025-07-22T21:20:08Z' }
  test-runtime: 1072.454
  threads:
  - thread: main
    messages:
    - { level: error, text: 'E> Child 200356 did not exit, sending signal SIGQUIT' }
```

Now:
```yaml
- test: selftest_freeze_exit_on_termination
  time-at-start: { elapsed:      0.000, now: !!timestamp '2025-07-23T18:02:32Z' }
  result: timed out
  fail: { cpu-mask: 'XXXX', time-to-fail: null, seed: 'LCG:1877619139'}
  time-at-end:   { elapsed:   1076.005, now: !!timestamp '2025-07-23T18:02:33Z' }
  test-runtime: 1076.064
  threads:
  - thread: main
    runtime: 1076.040
    resource-usage: { utime: 0.000, stime: 1.249, cpuavg: 0.1, maxrss: 3788, majflt: 0, minflt: 86, voluntary-cs: 12, involutary-cs: 8 }
    messages:
    - { level: error, text: 'E> Child 256147 did not exit, sending signal SIGQUIT' }
```

Selftests are currently Unix-only, because `TerminateProcess()` on
Windows does not allow interception. Windows tests only after #732.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>