FW/Logging: fix capturing of ps output on Linux for hung tests
Fixes commit f1d4860465f228bc9e60c856eeefc86018339bb5, which made this
use `std::string`: it opens a temporary file (memfd on Linux) for the
child process and then reads from it. But a temporary file is not a pipe
and the child process will sync its read/write position with the the
parent. Therefore, we must `lseek()` prior to reading from the fd.

And of course we must add the number of bytes read to `total_read`.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>