FW: re-disable stdout in debug builds
Updates 40f2b76248dd974948979a505fe66eac25702cd2, which allowed
`printf()` and family to print to stdout in debug builds, because
printf-debugging was useful with some libraries we had to use.

This retains that capability, but makes it an explicit opt-in, not
conditional in simply being a debug-mode build.

Drive-by remove the "no-triage" framework option.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>