Tests/x86: don't downgrade CPU target from what the user asked
The -Dmarch_base= config option sets a minimum CPU for the whole build.
So far, we've only used "westmere" (a.k.a. "x86-64-v2") and "haswell"
(a.k.a. "x86-64-v3"), which was ok. But if someone used a later version
like "x86-64-v4" or "diamondrapids", some or all of the tests would
downgrade to a lesser CPU.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>