FW/Test: preinit the tests in the main process
This is how it was meant to be, how it was designed back in day (commit
7c2b9cd8cc03828503da1526825dc5944ad1a3b6 in the internal repository
prior to open-sourcing), but was lost in refactoring with PR #495
(commit 7c2b9cd8cc03828503da1526825dc5944ad1a3b6).

This new design is stricter in that the preinit function can only affect
things that the parent process will see, for example, the timing and
looping constraints or replacing the init function (so long as it runs
in the parent).

No self-test for group inits because in the current design doesn't allow
the selftests to have group init, therefore we can't test the side-
effect.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>