FW: add flag `test_init_in_parent` to execute init() in the parent
This allows certain tests that always, deterministically and with little
overhead skip or fail to be run in the parent process, avoiding the need
to create a child process and waiting on it. This is going to be useful
for the placeholder tests that cannot be implemented in a given OS or
configuration, especially so for Windows where `CreateProcess` is rather
expensive.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>