FW/Logging: include the resource usage even for slices that didn't fail
We would get to `maybe_print_slice_resource_usage()` only if the main
thread had messages or had failed, which may not always be. This is
applicable for when the child crashes without writing anything and our
crash handler didn't run.

[ChangeLog][Logging] Updated the logging framework to include the
`resource-usage` entry (YAML output format) for crashing child
processes even when the crash handler didn't run, such as when
running with `--on-crash=core` or in the case of a SIGKILL signal.

Now:
```yaml
- test: selftest_sigkill
  details: { quality: production, description: "Raises SIGKILL" }
  state: { seed: 'AES:38a1dca1212b075a9ae16b2a9cda7347c75e235eded4f8a5651e94d563258cb8', iteration: 0, retry: false }
  time-at-start: { elapsed:      0.000, now: !!timestamp '2025-09-17T15:08:44Z' }
  result: crash
  result-details: { crashed: true, core-dump: false, code: 9, reason: 'Killed' }
  fail: { cpu-mask: null, time-to-fail: null, seed: 'AES:38a1dca1212b075a9ae16b2a9cda7347c75e235eded4f8a5651e94d563258cb8'}
  time-at-end:   { elapsed:     14.000, now: !!timestamp '2025-09-17T15:08:44Z' }
  test-runtime: 13.069
  threads:
  - thread: main
    runtime: 13.064
    resource-usage: { utime: 0.000, stime: 1.055, cpuavg: 8.1, maxrss: 5504, majflt: 0, minflt: 83, voluntary-cs: 11, involutary-cs: 7 }
    messages:
```

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>