FW/Win32: remove ucrt-mapping.c and the `_setjmp` redirect
This fixes the build with a newer MinGW runtime:
```
../framework/sysdeps/windows/ucrt-mapping.c:11:5: warning: '_setjmp' redeclared without dllimport attribute: previous dllimport ignored [-Wattributes]
   11 | int _setjmp(jmp_buf Buf, void *ctx)
      |     ^~~~~~~
../framework/sysdeps/windows/ucrt-mapping.c: In function '_setjmp':
../framework/sysdeps/windows/ucrt-mapping.c:13:12: error: implicit declaration of function '__intrinsic_setjmpex' [-Wimplicit-function-declaration]
   13 |     return __intrinsic_setjmpex(Buf, ctx);
      |            ^~~~~~~~~~~~~~~~~~~~
```

The error is new after an MSYS upgrade on my system. I don't know what
changed.

I also don't remember why this was needed. The old repository only
says:
```
commit dad439ed0adbdf2ca85217c24f033562d2a6907f
Author: Thiago Macieira <thiago.macieira@intel.com>
Date:   Wed Apr 7 19:51:42 2021 -0700

    Win32: add a file to map from old names to ucrt's
```

I'm guessing this was an experimental build, with libraries partly
built for MSVCRT and partly for UCRT. We didn't switch to UCRT until
May 2023.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>