FW/CPU: fix Linux dynamic XSAVE state requests
The `adjusted_xcr0()` function was designed when we thought new XSAVE
states would be allocated at the end of the last, thus allowing us to
make a single request for the highest numbered feature and thus get them
all. That turned out not to be true: APX-F reused the storage previously
allocated for MPX (a deprecated feature), so the Linux kernel requires
no request. More importantly, if you do request feature 19, it replies
with -EOPNOTSUPP (see [1]), which causes us to conclude the kernel
supports neither APX-F nor AMX.

So rewrite this to make a request for each of the bits that we require,
instead of trying to make a single request to get them all. Note that
the kernel only requires and accepts the request to XTILEDATA for AMX,
enabling both it and XTILECFG.

[1] https://codebrowser.dev/linux/linux/arch/x86/kernel/fpu/xstate.c.html#xstate_request_perm

Fixes: 2f9653845f ("FW/CPU: enable checking of XSAVE states")
Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>