FW/CPU-Topology: include the CPU type in the sorting plan
This will cause hybrid/heterogeneous systems to sort their CPUs by type
(inside of a package). I don't have any report of a system where this
will make a difference: for example, for the Intel Ultra Core Series 2
(Arrow Lake)[1], this review[2] indicates that the Lion Cove P-Cores and
the Skymont E-Cores are contiguous.

In order to have the field in the right position for the 16-byte sorting
block, I shrunk the `thread_id` to 8 bits only (there are never going to
be more than 127 threads in a core) and packed that portion of the
struct. We can't shrink `numa_id`: I have already seen systems with 64
sockets and 128 NUMA domains. I could remove the `tile_id` because it's
currently unused and should map 1:1 (or 2:1) to the NUMA ID.

[1] https://www.intel.com/content/www/us/en/ark/products/series/241071/intel-core-ultra-processors-series-2.html
[2] https://chipsandcheese.com/p/examining-intels-arrow-lake-at-the

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>