FW/CPU-Topology: Implement slicing per core type for hybrid CPUs
When running on systems with different core types, like Intel's
current line of client CPUs since the 12th Generation (Alder Lake), we
split the execution so the different core types run in different
slices. This should isolate the differences in allowed behaviour
(propagation of NaNs in the dot product instructions, which flags are
set as a result of IMUL, etc.) so that all cores in a given slice have
the same behaviour.

For example, on my ADL workstation:
```yaml
test-plans:
  fullsocket: [ { starting_cpu: 0, count: 20 } ]
  heuristic: [ { starting_cpu: 0, count: 4 }, { starting_cpu: 4, count: 16 } ]
```

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>