FW/CPU-Topology: add a way to mock CPU core type, module and NUMA IDs
This changes the way the `SANDSTONE_MOCK_TOPOLOGY` environment variable
is used so we can extend it to other fields in future. The format is
similar to that of the `--cpuset=` / `--deviceset=` command-line
parameters: 'p' for package, 'c' for core, etc.

This commit adds the ability to set the CPU core type (with prefix "y"),
module IDs ("m") and NUMA domain IDs ("n").

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>