CPU-separation: introduce `special_tests` vector to store `mce_test` & others
Non-CPU devices would be able to prepend this vector with its' own
device-specific tests expected to work the same way as `mce_check`.
That is, placed at the end of the test lists, and not being able to be
added manually (serving a purpose of a "monitor").
`mce_test` is still guaranteed to be executed as the very last one.

Signed-off-by: Magdalena Cisowska <magdalena.cisowska@intel.com>