FW: add flag `test_in_parent` to execute whole test in parent
This allows tests to be run in parent regardless of the command line
mode selected. It's a preparation step for device-specific special tests
that may want to spawn a 'monitoring' thread in their preinit (to listen for
driver events, for example) and join it in cleanup (test added at the end
of the test list).

Signed-off-by: Magdalena Cisowska <magdalena.cisowska@intel.com>