FW/Logging: add `memcmp_or_fail` with a callback
The callback can return a `std::string` to be printed in the
description. The callback can optionally take a `ptrdiff_t` first
parameter, which will be the index in the array where a difference was
observed.

The C version is called `memcmp_or_fail_cb` because the `memcmp_or_fail`
macro enforces a string as the first parameter. The callback signature
is also different: it returns a `char *` which will be freed using
`free()` and has a `void *` first parameter, to communicate the context.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>