FW/CPU-Topology: set defaults to mocked module_id and numa_id
This defaults `module_id` to the `core_id` if not provided, which
matches the way modules are usually detected when not provided.
Likewise, we are defaulting 1:1 packages (sockets) and NUMA domains.

Noticed that the CI was producing:
```
ok 116 slicing cores # skip Test only works with Debug builds (to mock the topology) or at least 2 different modules
```

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>