FW: add `test_schedule_isolate_numa_domain`
This flag allows tests to specify that they want to run in all
cores/modules of a given NUMA domain. This should map 1:1 to the tiles
in a high-core-count CPU, though I don't have access to one to show it
does.

For example, on an 120-core Granite Rapids, which has 3 NUMA domains of
40 cores each:
```yaml
test-plans:
  fullsocket: [ { starting_cpu: 0, count: 240 } ]
  isolate_numa: [ { starting_cpu: 0, count: 80 }, { starting_cpu: 80, count: 80 }, { starting_cpu: 160, count: 80 } ]
  heuristic: [ { starting_cpu: 0, count: 40 }, { starting_cpu: 40, count: 40 }, { starting_cpu: 80, count: 40 }, { starting_cpu: 120, count: 40 }, { starting_cpu: 160, count: 40 }, { starting_cpu: 200, count: 40 } ]
```

As a side-effect of the way I implemented the core-type separation in
PR#853, each core type is considered a separate "NUMA" domain, so on my
Alder Lake workstation I get:
```yaml
test-plans:
  fullsocket: [ { starting_cpu: 0, count: 20 } ]
  isolate_numa: [ { starting_cpu: 0, count: 16 }, { starting_cpu: 16, count: 4 } ]
  heuristic: [ { starting_cpu: 0, count: 16 }, { starting_cpu: 16, count: 4 } ]
```

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>