CI/Bats: fix the "slicing core types" test on new non-hybrid systems
It is possible for homogeneous systems to still report core type in the
CPUID leaf, which Intel's Xeon6P (Granite Rapids) do. Fix the test to
work there and not assume that any system that reports a core type must
have both types of cores.

I could have fixed the test to only run on hybrid parts (or in debug
builds), but this is a more interesting test.

```yaml
cpu-info:
- { logical:   0, package: 0, numa_node: 0, module:   0, core:   0, thread: 0, core_type: p, family: 6, model: 0xad, stepping: 1, microcode: 0x110003d0, ppin: null }   # 0
- { logical: 192, package: 0, numa_node: 0, module:   0, core:   0, thread: 1, core_type: p, family: 6, model: 0xad, stepping: 1, microcode: 0x110003d0, ppin: null }   # 1
- { logical:   1, package: 0, numa_node: 0, module:   1, core:   1, thread: 0, core_type: p, family: 6, model: 0xad, stepping: 1, microcode: 0x110003d0, ppin: null }   # 2
...
test-plans:
  fullsocket: [ { starting_cpu: 0, count: 192 } ]
  heuristic: [ { starting_cpu: 0, count: 192 } ]
```

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>