FW/Logging: mark `logging_mark_thread_{failed,skipped}()` noexcept
The `logging_mark_thread_failed()` function is called from the Unix
child_debug signal handler, so it already has a comment indicating it
must be async-signal-safe. In theory, async-signal-safe functions *can*
throw and even unwind to a catch point outside the handler, but we're
not using that.

So help codegen by with the marker.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>