FW/Logging: add support for installing a callback after errors
Many tests would like to print some extra, useful information to help in
debugging or identifying problems. For example, accelerator tests need
to print which accelerator was being tested (at least until the
framework is taught to target specific accelerators). But we don't want
to print any information *unless* the test fails, to avoid cluttering
logs.

This commit add supports for installing a callback, which the framework
will call after the first error in any given thread. It's the
responsibility of the test to retain a valid state throughout, including
if the test may `return EXIT_FAILURE;`. The `log_yaml()` function (PR #869)
comes in very handy for this.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>