FW: do not call `restrict_topology()` inside `commit_shmem()`
All we need from this call is to update `cpu_info` ptr, as `sApp->shmem`
may have moved. No need to call build_topology() at this point, since device range has not changed.
The way it's been implemented was enforcing other implementations of `restrict_topology()` to
know that they must provide update of the pointer inside of it as well. Since this step is
completely device-agnostic, let's just call it directly at the end of `commit_shmem()`.

Signed-off-by: Magdalena Cisowska <magdalena.cisowska@intel.com>